swapt_site/requirementa.txt
asgiref==3.4.1; python_version >= '3.6'
django==3.2.5
djangorestframework==3.12.4
pytz==2021.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.1; python_version >= '3.5'